+++ presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractMapBlock.java
@@ -141,7 +141,7 @@ public abstract class AbstractMapBlock
     }
 
     @Override
-    public int getRegionSizeInBytes(int position, int length)
+    public long getRegionSizeInBytes(int position, int length)
     {
         int positionCount = getPositionCount();
         if (position < 0 || length < 0 || position + length > positionCount) {
